+++ src/common/utils/index.ts
@@ -10,11 +10,12 @@ import * as util from 'util';
 import * as xml2js from 'xml2js';
 const exec = util.promisify(child.exec);
 import { SfdxError } from '@salesforce/core';
-import simpleGit, { SimpleGit } from 'simple-git';
+import ora = require('ora');
+import simpleGit, { FileStatusResult, SimpleGit } from 'simple-git';
 import { CONSTANTS } from '../../config';
 import { MetadataUtils } from '../metadata-utils';
 
-let git: SimpleGit = null;
+export let git: SimpleGit = null;
 
 if (isGitRepo()) {
   initSimpleGit();
@@ -123,6 +124,11 @@ export async function getCurrentGitBranch(options: any = { formatted: false }) {
   return gitBranch;
 }
 
+export async function gitCheckOutRemote(branchName: string) {
+  await git.checkout(branchName);
+  await git.pull();
+}
+
 // Get local git branch name
 export async function ensureGitBranch(branchName: string, options: any = { init: false }) {
   if (git == null) {
@@ -145,6 +151,68 @@ export async function ensureGitBranch(branchName: string, options: any = { init:
     }
   }
   return true;
+}
+
+// Checks that current git status is clean.
+export async function checkGitClean(options: any) {
+  if (git == null) {
+    throw new SfdxError('[sfdx-hardis] You must be within a git repository');
+  }
+  const gitStatus = await git.status();
+  if (gitStatus.files.length > 0) {
+    const localUpdates = gitStatus.files.map((fileStatus: FileStatusResult) => {
+      return `(${fileStatus.working_dir}) ${fileStatus.path}`;
+    }).join('\n');
+    throw new SfdxError(`[sfdx-hardis] Branch ${c.bold(gitStatus.current)} is not clean. You must ${c.bold('commit or cancel')} the following local updates:\n${c.yellow(localUpdates)}`);
+  }
+}
+
+// Interactive git add
+export async function interactiveGitAdd(options: any = { filter: [], unstage: true }) {
+  if (git == null) {
+    throw new SfdxError('[sfdx-hardis] You must be within a git repository');
+  }
+  const gitStatus = await git.status();
+  const filesFiltered = gitStatus.files.filter((fileStatus: FileStatusResult) => {
+    return options.filter.filter((filterString: string) => fileStatus.path.includes(filterString)).length === 0;
+  });
+  let listOfFiles = [];
+  if (filesFiltered.length > 0) {
+    const selectFilesStatus = await prompts({
+      type: 'multiselect',
+      name: 'files',
+      message: c.cyanBright('Please select the files you want to commit (save)'),
+      choices: filesFiltered.map((fileStatus: FileStatusResult) => {
+        return { title: `(${fileStatus.working_dir}) ${fileStatus.path}`, value: fileStatus };
+      })
+    });
+    const listOfFilesDisplay = selectFilesStatus.files.map((fileStatus: FileStatusResult) => {
+      return `(${fileStatus.working_dir}) ${fileStatus.path}`;
+    }).join('\n');
+    const commitFilesResponse = await prompts({
+      type: 'confirm',
+      name: 'commit',
+      message: c.cyanBright(`Do you confirm that you want to commit the following list of files ?\n${c.yellow(listOfFilesDisplay)}`),
+      choices: filesFiltered.map((fileStatus: FileStatusResult) => {
+        return { title: `(${fileStatus.working_dir}) ${fileStatus.path}`, value: fileStatus };
+      })
+    });
+    listOfFiles = selectFilesStatus.files.map((fileStatus: FileStatusResult) => {
+      if (fileStatus.path.startsWith('"')) {
+        fileStatus.path = fileStatus.path.substring(1);
+      }
+      if (fileStatus.path.endsWith('"')) {
+        fileStatus.path = fileStatus.path.slice(0, -1);
+      }
+      return fileStatus.path;
+    });
+    if (commitFilesResponse.commit === true) {
+      await git.add(listOfFiles);
+    }
+  } else {
+    uxLog(this, c.cyan('No uncommited local files'));
+  }
+  return listOfFiles;
 }
 
 // Shortcut to add, commit and push
@@ -192,17 +260,24 @@ export async function execCommand(
   options: any = {
     fail: false,
     output: false,
-    debug: false
+    debug: false,
+    spinner: true
   }
 ): Promise<any> {
-  uxLog(commandThis, `[sfdx-hardis][command] ${c.bold(c.grey(command))}`);
+  const commandLog = `[sfdx-hardis][command] ${c.bold(c.grey(command))}`;
   let commandResult = null;
   // Call command (disable color before for json parsing)
   const prevForceColor = process.env.FORCE_COLOR;
   process.env.FORCE_COLOR = '0';
+  const spinner = ora({text: commandLog, spinner: 'moon'}).start();
+  if (options.spinner === false) {
+    spinner.stop();
+  }
   try {
     commandResult = await exec(command, { maxBuffer: 10000 * 10000 });
+    spinner.succeed();
   } catch (e) {
+    spinner.fail();
     process.env.FORCE_COLOR = prevForceColor;
     // Display error in red if not json
     if (!command.includes('--json') || options.fail) {
@@ -235,12 +310,15 @@ export async function execCommand(
         c.red(`[sfdx-hardis][ERROR] Command failed: ${commandResult}`)
       );
     }
+    if (commandResult.stderr && commandResult.stderr.length > 2) {
+      uxLog(this, '[sfdx-hardis][WARNING] stderr: ' + c.yellow(commandResult.stderr));
+    }
     return parsedResult;
   } catch (e) {
     // Manage case when json is not parseable
     return {
       status: 1,
-      errorMessage: `[sfdx-hardis][ERROR] Error parsing JSON in command result: ${e.message}\n${commandResult.stdout}`
+      errorMessage: c.red(`[sfdx-hardis][ERROR] Error parsing JSON in command result: ${e.message}\n${commandResult.stdout}\n${commandResult.stderr})`)
     };
   }
 }
@@ -491,7 +569,7 @@ export async function generateSSLCertificate(branchName: string, folder: string,
   await execCommand(`openssl genrsa -des3 -passout "pass:${pwd}" -out server.pass.key 2048`, this, { output: true, fail: true });
   await execCommand(`openssl rsa -passin "pass:${pwd}" -in server.pass.key -out server.key`, this, { output: true, fail: true });
   await fs.remove('server.pass.key');
-  await prompts({type: 'confirm', message: c.cyanBright('Now answer the following questions. The answers are not really important :)\nHit ENTER when ready')});
+  await prompts({ type: 'confirm', message: c.cyanBright('Now answer the following questions. The answers are not really important :)\nHit ENTER when ready') });
   await new Promise((resolve, reject) => {
     const opensslCommand = 'openssl req -new -key server.key -out server.csr';
     crossSpawn(opensslCommand, [], { stdio: 'inherit' }).on('close', () => {